5G_Emulator_API/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
opentelemetry-exporter-prometheus
opentelemetry-exporter-console
fastapi
requests
psutil
prometheus_client
uvicorn
